+++ src/components/__init__.py
@@ -5,3 +5,4 @@ from components.form import Form
 from components.model_selector import ModelSelector
 from components.llm_input import LLMInput
 from components.llm_list import LLMList
+from components.evaluation_visualizer import EvaluationVisualizer